+++ lib/shared/src/mentions/api.ts
@@ -1,9 +1,5 @@
-import type { ContextItem, ContextItemWithContent } from '../codebase-context/messages'
-import type { Configuration } from '../configuration'
 import { openCtx } from '../context/openctx/api'
 import { logDebug } from '../logger'
-import type { PromptString } from '../prompt/prompt-string'
-import { URL_CONTEXT_MENTION_PROVIDER } from './providers/urlMentions'
 
 /**
  * A unique identifier for a {@link ContextMentionProvider}.
@@ -11,20 +7,17 @@ import { URL_CONTEXT_MENTION_PROVIDER } from './providers/urlMentions'
 export type ContextMentionProviderID = string
 
 /**
- * Providers that supply context that the user can @-mention in chat.
- *
- * This API is *experimental* and subject to rapid, unannounced change.
- *
+ * Props required by context item providers to return possible context items.
  */
-export const CONTEXT_MENTION_PROVIDERS: ContextMentionProvider[] = [URL_CONTEXT_MENTION_PROVIDER]
+export interface ContextItemProps {
+    gitRemotes: { hostname: string; owner: string; repoName: string; url: string }[]
+}
 
 /**
- * A provider that can supply context for users to @-mention in chat.
- *
- * This API is *experimental* and subject to rapid, unannounced change.
+ * Metadata about a {@link ContextMentionProvider}.
  */
-export interface ContextMentionProvider<ID extends ContextMentionProviderID = ContextMentionProviderID> {
-    id: ID
+export interface ContextMentionProviderMetadata {
+    id: string
 
     /**
      * A short, human-readable display title for the provider, such as "Google Docs". If not given,
@@ -41,60 +34,16 @@ export interface ContextMentionProvider<ID extends ContextMentionProviderID = Co
      * Human-readable display string for when the provider has no items for the query.
      */
     emptyLabel?: string
-
-    /**
-     * Get a list of possible context items to show (in a completion menu) when the user triggers
-     * this provider while typing `@` in a chat message.
-     *
-     * {@link query} omits the `@`.
-     */
-    queryContextItems(
-        query: string,
-        props: ContextItemProps,
-        signal?: AbortSignal
-    ): Promise<ContextItemFromProvider<ID>[]>
-
-    /**
-     * Resolve a context item to one or more items that have the {@link ContextItem.content} field
-     * filled in. A provider is called to resolve only the context items that it returned in
-     * {@link queryContextItems} and that the user explicitly added.
-     */
-    resolveContextItem?(
-        item: ContextItemFromProvider<ID>,
-        input: PromptString,
-        signal?: AbortSignal
-    ): Promise<ContextItemWithContent[]>
 }
 
-/**
- * Props required by context item providers to return possible context items.
- */
-export interface ContextItemProps {
-    gitRemotes: { hostname: string; owner: string; repoName: string; url: string }[]
-}
-
-export type ContextItemFromProvider<ID extends ContextMentionProviderID> = ContextItem & {
-    /**
-     * The ID of the {@link ContextMentionProvider} that supplied this context item.
-     */
-    provider: ID
-}
-
-/**
- * Metadata about a {@link ContextMentionProvider}.
- */
-export interface ContextMentionProviderMetadata<
-    ID extends ContextMentionProviderID = ContextMentionProviderID,
-> extends Pick<ContextMentionProvider<ID>, 'id' | 'title' | 'queryLabel' | 'emptyLabel'> {}
-
-export const FILE_CONTEXT_MENTION_PROVIDER: ContextMentionProviderMetadata<'file'> = {
+export const FILE_CONTEXT_MENTION_PROVIDER: ContextMentionProviderMetadata & { id: 'file' } = {
     id: 'file',
     title: 'Files',
     queryLabel: 'Search for a file to include...',
     emptyLabel: 'No files found',
 }
 
-export const SYMBOL_CONTEXT_MENTION_PROVIDER: ContextMentionProviderMetadata<'symbol'> = {
+export const SYMBOL_CONTEXT_MENTION_PROVIDER: ContextMentionProviderMetadata & { id: 'symbol' } = {
     id: 'symbol',
     title: 'Symbols',
     queryLabel: 'Search for a symbol to include...',
@@ -102,9 +51,7 @@ export const SYMBOL_CONTEXT_MENTION_PROVIDER: ContextMentionProviderMetadata<'sy
 }
 
 /** Metadata for all registered {@link ContextMentionProvider}s. */
-export async function allMentionProvidersMetadata(
-    config: Pick<Configuration, 'experimentalNoodle' | 'experimentalURLContext'>
-): Promise<ContextMentionProviderMetadata[]> {
+export async function allMentionProvidersMetadata(): Promise<ContextMentionProviderMetadata[]> {
     const items = [
         FILE_CONTEXT_MENTION_PROVIDER,
         SYMBOL_CONTEXT_MENTION_PROVIDER,
@@ -129,7 +76,7 @@ export async function openCtxMentionProviders(): Promise<ContextMentionProviderM
                 id: provider.providerUri,
                 title: provider.name,
                 queryLabel: provider.name,
-                emptyLabel: 'No results found',
+                emptyLabel: 'No results',
             }))
             .sort((a, b) => (a.title > b.title ? 1 : -1))
     } catch (error) {