+++ vscode/src/main.ts
@@ -2,18 +2,17 @@ import * as vscode from 'vscode'
 
 import {
     type AuthStatus,
+    type ChatClient,
     ConfigFeaturesSingleton,
     type ConfigurationWithAccessToken,
     type DefaultCodyCommands,
-    type EventSource,
+    type Guardrails,
     ModelsService,
     PromptMixin,
     PromptString,
     contextFiltersProvider,
     featureFlagProvider,
-    githubClient,
     graphqlClient,
-    newPromptMixin,
     setClientNameVersion,
     setLogger,
     telemetryRecorder,
@@ -46,7 +45,8 @@ import type { CodyCommandArgs } from './commands/types'
 import { newCodyCommandArgs } from './commands/utils/get-commands'
 import { createInlineCompletionItemProvider } from './completions/create-inline-completion-item-provider'
 import { createInlineCompletionItemFromMultipleProviders } from './completions/create-multi-model-inline-completion-provider'
-import { getConfiguration, getFullConfig } from './configuration'
+import { getFullConfig } from './configuration'
+import { BaseConfigWatcher, type ConfigWatcher } from './configwatcher'
 import { EnterpriseContextFactory } from './context/enterprise-context-factory'
 import { exposeOpenCtxClient } from './context/openctx'
 import { EditManager } from './edit/manager'
@@ -55,19 +55,20 @@ import { VSCodeEditor } from './editor/vscode-editor'
 import type { PlatformContext } from './extension.common'
 import { configureExternalServices } from './external-services'
 import { isRunningInsideAgent } from './jsonrpc/isRunningInsideAgent'
+import type { ContextRankingController } from './local-context/context-ranking'
+import type { LocalEmbeddingsController } from './local-context/local-embeddings'
+import type { SymfRunner } from './local-context/symf'
 import { logDebug, logError } from './log'
 import { MinionOrchestrator } from './minion/MinionOrchestrator'
 import { PoorMansBash } from './minion/environment'
-import { getChatModelsFromConfiguration, syncModels } from './models/sync'
+import { registerModelsFromVSCodeConfiguration, syncModels } from './models/sync'
 import { CodyProExpirationNotifications } from './notifications/cody-pro-expiration'
 import { showSetupNotification } from './notifications/setup-notification'
 import { initVSCodeGitApi } from './repository/git-extension-api'
 import { repoNameResolver } from './repository/repo-name-resolver'
-import { SearchViewProvider } from './search/SearchViewProvider'
 import { AuthProvider } from './services/AuthProvider'
 import { CharactersLogger } from './services/CharactersLogger'
 import { showFeedbackSupportQuickPick } from './services/FeedbackOptions'
-import { GuardrailsProvider } from './services/GuardrailsProvider'
 import { displayHistoryQuickPick } from './services/HistoryChat'
 import { localStorage } from './services/LocalStorageProvider'
 import { VSCodeSecretStorage, getAccessToken, secretStorage } from './services/SecretStorageProvider'
@@ -105,27 +106,15 @@ export async function start(
 
     const disposables: vscode.Disposable[] = []
 
-    const { disposable, onConfigurationChange } = await register(
-        context,
-        await getFullConfig(),
-        platform
-    )
-    disposables.push(disposable)
+    const configWatcher = await BaseConfigWatcher.create(getFullConfig, disposables)
 
-    // Re-initialize when configuration
-    disposables.push(
-        vscode.workspace.onDidChangeConfiguration(async event => {
-            if (event.affectsConfiguration('cody')) {
-                const config = await getFullConfig()
-                await onConfigurationChange(config)
-                platform.onConfigurationChange?.(config)
-                if (config.chatPreInstruction.length > 0) {
-                    PromptMixin.addCustom(newPromptMixin(config.chatPreInstruction))
-                }
-                getChatModelsFromConfiguration()
-            }
-        })
-    )
+    configWatcher.onChange(async config => {
+        platform.onConfigurationChange?.(config)
+        registerModelsFromVSCodeConfiguration()
+    }, disposables)
+
+    const { disposable } = await register(context, configWatcher, platform)
+    disposables.push(disposable)
 
     return vscode.Disposable.from(...disposables)
 }
@@ -133,43 +122,38 @@ export async function start(
 // Registers commands and webview given the config.
 const register = async (
     context: vscode.ExtensionContext,
-    initialConfig: ConfigurationWithAccessToken,
+    configWatcher: ConfigWatcher<ConfigurationWithAccessToken>,
     platform: PlatformContext
 ): Promise<{
     disposable: vscode.Disposable
-    onConfigurationChange: (newConfig: ConfigurationWithAccessToken) => Promise<void>
 }> => {
-    setClientNameVersion(platform.extensionClient.clientName, platform.extensionClient.clientVersion)
-    const authProvider = AuthProvider.create(initialConfig)
-    await localStorage.setConfig(initialConfig)
-
     const disposables: vscode.Disposable[] = []
-    // Initialize `displayPath` first because it might be used to display paths in error messages
-    // from the subsequent initialization.
-    disposables.push(manageDisplayPathEnvInfoForExtension())
-
-    disposables.push(await initVSCodeGitApi())
-
+    const initialConfig = configWatcher.get()
     const isExtensionModeDevOrTest =
         context.extensionMode === vscode.ExtensionMode.Development ||
         context.extensionMode === vscode.ExtensionMode.Test
 
-    await configureEventsInfra(initialConfig, isExtensionModeDevOrTest, authProvider)
+    setClientNameVersion(platform.extensionClient.clientName, platform.extensionClient.clientVersion)
+    const authProvider = AuthProvider.create(initialConfig)
 
-    const editor = new VSCodeEditor()
+    // Initialize `displayPath` first because it might be used to display paths in error messages
+    // from the subsequent initialization.
+    disposables.push(manageDisplayPathEnvInfoForExtension())
 
-    // Could we use the `initialConfig` instead?
-    const workspaceConfig = vscode.workspace.getConfiguration()
-    const config = getConfiguration(workspaceConfig)
+    // Init local storage
+    await configWatcher.initAndOnChange(async config => {
+        await localStorage.setConfig(config)
+    }, disposables)
 
-    if (config.chatPreInstruction.length > 0) {
-        PromptMixin.addCustom(newPromptMixin(config.chatPreInstruction))
-    }
+    // Ensure Git API is available
+    disposables.push(await initVSCodeGitApi())
 
-    void parseAllVisibleDocuments()
+    // Telemetry
+    await configWatcher.initAndOnChange(async config => {
+        await configureEventsInfra(config, isExtensionModeDevOrTest, authProvider)
+    }, disposables)
 
-    disposables.push(vscode.window.onDidChangeVisibleTextEditors(parseAllVisibleDocuments))
-    disposables.push(vscode.workspace.onDidChangeTextDocument(updateParseTreeOnEdit))
+    registerParserListeners(disposables)
 
     // Enable tracking for pasting chat responses into editor text
     disposables.push(
@@ -185,11 +169,15 @@ const register = async (
 
     await authProvider.init()
 
-    exposeOpenCtxClient(context.secrets, initialConfig)
-    graphqlClient.onConfigurationChange(initialConfig)
-    githubClient.onConfigurationChange({ authToken: initialConfig.experimentalGithubAccessToken })
-    void featureFlagProvider.syncAuthStatus()
+    await exposeOpenCtxClient(context.secrets, initialConfig)
 
+    await configWatcher.initAndOnChange(async config => {
+        graphqlClient.onConfigurationChange(config)
+        // githubClient.onConfigurationChange({ authToken: config.experimentalGithubAccessToken })
+        await featureFlagProvider.syncAuthStatus()
+    }, disposables)
+
+    // Initialize external services
     const {
         chatClient,
         completionsClient,
@@ -200,19 +188,20 @@ const register = async (
         onConfigurationChange: externalServicesOnDidConfigurationChange,
         symfRunner,
     } = await configureExternalServices(context, initialConfig, platform, authProvider)
+    configWatcher.onChange(async config => {
+        externalServicesOnDidConfigurationChange(config)
+        symfRunner?.setSourcegraphAuth(config.serverEndpoint, config.accessToken)
+    }, disposables)
 
     if (symfRunner) {
         disposables.push(symfRunner)
     }
 
-    //
-    // Minion stuff
-    //
-    if (config.experimentalMinionAnthropicKey) {
+    // Initialize Minion
+    if (initialConfig.experimentalMinionAnthropicKey) {
         const minionOrchestrator = new MinionOrchestrator(context.extensionUri, authProvider, symfRunner)
         disposables.push(minionOrchestrator)
         disposables.push(
-            // Minion
             vscode.commands.registerCommand('cody.minion.panel.new', () =>
                 minionOrchestrator.createNewMinionPanel()
             ),
@@ -223,9 +212,15 @@ const register = async (
         )
     }
 
+    // Initialize enterprise context
     const enterpriseContextFactory = new EnterpriseContextFactory(completionsClient)
     disposables.push(enterpriseContextFactory)
+    configWatcher.onChange(async () => {
+        enterpriseContextFactory.clientConfigurationDidChange()
+    }, disposables)
 
+    // Initialize context provider
+    const editor = new VSCodeEditor()
     const contextProvider = new ContextProvider(
         initialConfig,
         editor,
@@ -233,126 +228,53 @@ const register = async (
         localEmbeddings,
         enterpriseContextFactory.createRemoteSearch()
     )
-    disposables.push(contextFiltersProvider)
     disposables.push(contextProvider)
+    disposables.push(contextFiltersProvider)
     await contextFiltersProvider
         .init(repoNameResolver.getRepoNamesFromWorkspaceUri)
         .then(() => contextProvider.init())
 
-    // Shared configuration that is required for chat views to send and receive messages
-    const messageProviderOptions: MessageProviderOptions = {
-        chat: chatClient,
-        guardrails,
-        editor,
-        authProvider,
-        contextProvider,
-    }
+    configWatcher.onChange(async config => {
+        await contextFiltersProvider.init(repoNameResolver.getRepoNamesFromWorkspaceUri)
+        await contextProvider.onConfigurationChange(config)
+        await localEmbeddings?.setAccessToken(config.serverEndpoint, config.accessToken)
+    }, disposables)
 
-    const chatManager = new ChatManager(
+    const { chatManager, editorManager } = registerChat(
         {
-            ...messageProviderOptions,
-            extensionUri: context.extensionUri,
-            config,
-            startTokenReceiver: platform.startTokenReceiver,
+            context,
+            configWatcher,
+            platform,
+            chatClient,
+            guardrails,
+            editor,
+            authProvider,
+            contextProvider,
+            enterpriseContextFactory,
+            localEmbeddings,
+            contextRanking,
+            symfRunner,
         },
-        chatClient,
-        enterpriseContextFactory,
-        localEmbeddings || null,
-        contextRanking || null,
-        symfRunner || null,
-        guardrails
+        disposables
     )
 
-    const ghostHintDecorator = new GhostHintDecorator(authProvider)
-    const editorManager = new EditManager({
-        chat: chatClient,
-        editor,
-        ghostHintDecorator,
-        authProvider,
-        extensionClient: platform.extensionClient,
-    })
-    disposables.push(ghostHintDecorator, editorManager, new CodeActionProvider({ contextProvider }))
-
-    let oldConfig = JSON.stringify(initialConfig)
-    async function onConfigurationChange(newConfig: ConfigurationWithAccessToken): Promise<void> {
-        if (oldConfig === JSON.stringify(newConfig)) {
-            return Promise.resolve()
-        }
-
-        ModelsService.onConfigChange(newConfig.experimentalOllamaChat)
-
-        const promises: Promise<void>[] = []
-        oldConfig = JSON.stringify(newConfig)
-
-        promises.push(featureFlagProvider.syncAuthStatus())
-        graphqlClient.onConfigurationChange(newConfig)
-        exposeOpenCtxClient(secretStorage, newConfig)
-        upstreamHealthProvider.onConfigurationChange(newConfig)
-        githubClient.onConfigurationChange({ authToken: initialConfig.experimentalGithubAccessToken })
-        promises.push(
-            contextFiltersProvider
-                .init(repoNameResolver.getRepoNamesFromWorkspaceUri)
-                .then(() => contextProvider.onConfigurationChange(newConfig))
-        )
-        externalServicesOnDidConfigurationChange(newConfig)
-        promises.push(configureEventsInfra(newConfig, isExtensionModeDevOrTest, authProvider))
-        platform.onConfigurationChange?.(newConfig)
-        symfRunner?.setSourcegraphAuth(newConfig.serverEndpoint, newConfig.accessToken)
-        enterpriseContextFactory.clientConfigurationDidChange()
-        promises.push(
-            localEmbeddings?.setAccessToken(newConfig.serverEndpoint, newConfig.accessToken) ??
-                Promise.resolve()
-        )
-        promises.push(setupAutocomplete())
-        promises.push(localStorage.setConfig(newConfig))
-        await Promise.all(promises)
-    }
-
-    // Register tree views
-    disposables.push(
-        chatManager,
-        vscode.window.registerWebviewViewProvider('cody.chat', chatManager.sidebarViewController, {
-            webviewOptions: { retainContextWhenHidden: true },
-        }),
-        // Update external services when configurationChangeEvent is fired by chatProvider
-        contextProvider.configurationChangeEvent.event(async () => {
-            const newConfig = await getFullConfig()
-            await onConfigurationChange(newConfig)
-        })
-    )
-
-    const statusBar = createStatusBar()
     const sourceControl = new CodySourceControl(chatClient)
-
-    // Important to respect `config.experimentalSymfContext`. The agent
-    // currently crashes with a cryptic error when running with symf enabled so
-    // we need a way to reliably disable symf until we fix the root problem.
-    if (symfRunner && config.experimentalSymfContext) {
-        const searchViewProvider = new SearchViewProvider(symfRunner)
-        disposables.push(searchViewProvider)
-        searchViewProvider.initialize()
-    }
-
-    if (localEmbeddings) {
-        // kick-off embeddings initialization
-        localEmbeddings.start()
-    }
-
-    if (config.experimentalSupercompletions) {
-        disposables.push(new SupercompletionProvider({ statusBar, chat: chatClient }))
-    }
+    const statusBar = createStatusBar()
 
     // Adds a change listener to the auth provider that syncs the auth status
     authProvider.addChangeListener(async (authStatus: AuthStatus) => {
-        syncModels(authStatus)
+        // Reset the available models based on the auth change.
+        await syncModels(authStatus)
+
         // Chat Manager uses Simple Context Provider
         await chatManager.syncAuthStatus(authStatus)
         editorManager.syncAuthStatus(authStatus)
         // Update context provider first it will also update the configuration
         await contextProvider.syncAuthStatus()
+
         const parallelPromises: Promise<void>[] = []
-        parallelPromises.push(featureFlagProvider.syncAuthStatus())
         // feature flag provider
+        parallelPromises.push(featureFlagProvider.syncAuthStatus())
         // Symf
         if (symfRunner && authStatus.isLoggedIn) {
             parallelPromises.push(
@@ -365,18 +287,31 @@ const register = async (
         }
         parallelPromises.push(setupAutocomplete())
         await Promise.all(parallelPromises)
+
         statusBar.syncAuthStatus(authStatus)
         sourceControl.syncAuthStatus(authStatus)
+
+        // Set the default prompt mixin on auth status change.
+        await PromptMixin.updateContextPreamble(isExtensionModeDevOrTest || isRunningInsideAgent())
     })
+
+    if (initialConfig.experimentalSupercompletions) {
+        disposables.push(new SupercompletionProvider({ statusBar, chat: chatClient }))
+    }
+
+    configWatcher.onChange(async () => {
+        setupAutocomplete()
+    }, disposables)
 
     // Sync initial auth status
     const initAuthStatus = authProvider.getAuthStatus()
-    syncModels(initAuthStatus)
+    await syncModels(initAuthStatus)
     await chatManager.syncAuthStatus(initAuthStatus)
     editorManager.syncAuthStatus(initAuthStatus)
-    ModelsService.onConfigChange(initialConfig.experimentalOllamaChat)
+    await configWatcher.initAndOnChange(() => ModelsService.onConfigChange(), disposables)
     statusBar.syncAuthStatus(initAuthStatus)
     sourceControl.syncAuthStatus(initAuthStatus)
+    await PromptMixin.updateContextPreamble(isExtensionModeDevOrTest || isRunningInsideAgent())
 
     const commandsManager = platform.createCommandsProvider?.()
     setCommandController(commandsManager)
@@ -486,9 +421,6 @@ const register = async (
             }
         ),
         // Chat
-        vscode.commands.registerCommand('cody.focus', () =>
-            vscode.commands.executeCommand('cody.chat.focus')
-        ),
         vscode.commands.registerCommand('cody.settings.extension', () =>
             vscode.commands.executeCommand('workbench.action.openSettings', {
                 query: '@ext:sourcegraph.cody-ai',
@@ -554,7 +486,7 @@ const register = async (
                 if (uri.path === '/app-done') {
                     // This is an old re-entrypoint from App that is a no-op now.
                 } else {
-                    await authProvider.tokenCallbackHandler(uri, config.customHeaders)
+                    await authProvider.tokenCallbackHandler(uri, initialConfig.customHeaders)
                 }
             },
         }),
@@ -578,31 +510,6 @@ const register = async (
                 'sourcegraph.cody-ai#welcome',
                 false
             )
-        }),
-        vscode.commands.registerCommand('cody.welcome-mock', () =>
-            vscode.commands.executeCommand(
-                'workbench.action.openWalkthrough',
-                'sourcegraph.cody-ai#welcome',
-                false
-            )
-        ),
-        vscode.commands.registerCommand('cody.walkthrough.showLogin', () =>
-            vscode.commands.executeCommand('workbench.view.extension.cody')
-        ),
-        vscode.commands.registerCommand('cody.walkthrough.showChat', () =>
-            chatManager.setWebviewView('chat')
-        ),
-        vscode.commands.registerCommand('cody.walkthrough.showFixup', () =>
-            chatManager.setWebviewView('chat')
-        ),
-        vscode.commands.registerCommand('cody.walkthrough.showExplain', async () => {
-            telemetryService.log(
-                'CodyVSCodeExtension:walkthrough:clicked',
-                { page: 'showExplain' },
-                { hasV2Event: true }
-            )
-            telemetryRecorder.recordEvent('cody.walkthrough.showExplain', 'clicked')
-            await chatManager.setWebviewView('chat')
         }),
 
         // StatusBar Commands
@@ -634,17 +541,6 @@ const register = async (
             vscode.window.showInformationMessage,
             vscode.env.openExternal
         ),
-        // For register sidebar clicks
-        vscode.commands.registerCommand('cody.sidebar.click', (name: string, command: string) => {
-            const source: EventSource = 'sidebar'
-            telemetryService.log(`CodyVSCodeExtension:command:${name}:clicked`, {
-                source,
-            })
-            telemetryRecorder.recordEvent(`cody.command.${name}`, 'clicked', {
-                privateMetadata: { source },
-            })
-            void vscode.commands.executeCommand(command, [source])
-        }),
         ...setUpCodyIgnore(initialConfig),
         // For debugging
         vscode.commands.registerCommand('cody.debug.export.logs', () => exportOutputLog(context.logUri)),
@@ -652,8 +548,11 @@ const register = async (
         vscode.commands.registerCommand('cody.debug.enable.all', () => enableVerboseDebugMode()),
         vscode.commands.registerCommand('cody.debug.reportIssue', () => openCodyIssueReporter()),
         new CharactersLogger(),
-        upstreamHealthProvider.onConfigurationChange(initialConfig)
+        upstreamHealthProvider
     )
+    await configWatcher.initAndOnChange(async config => {
+        upstreamHealthProvider.onConfigurationChange(config)
+    }, disposables)
 
     let setupAutocompleteQueue = Promise.resolve() // Create a promise chain to avoid parallel execution
 
@@ -705,9 +604,7 @@ const register = async (
                         client: codeCompletionsClient,
                         statusBar,
                         authProvider,
-                        triggerNotice: notice => {
-                            void chatManager.triggerNotice(notice)
-                        },
+
                         createBfgRetriever: platform.createBfgRetriever,
                     })
                 )
@@ -717,9 +614,7 @@ const register = async (
                         client: codeCompletionsClient,
                         statusBar,
                         authProvider,
-                        triggerNotice: notice => {
-                            void chatManager.triggerNotice(notice)
-                        },
+
                         createBfgRetriever: platform.createBfgRetriever,
                     })
                 )
@@ -731,15 +626,6 @@ const register = async (
     }
 
     const autocompleteSetup = setupAutocomplete().catch(() => {})
-
-    if (initialConfig.experimentalGuardrails) {
-        const guardrailsProvider = new GuardrailsProvider(guardrails, editor)
-        disposables.push(
-            vscode.commands.registerCommand('cody.guardrails.debug', async () => {
-                await guardrailsProvider.debugEditorSelection()
-            })
-        )
-    }
 
     if (!isRunningInsideAgent()) {
         // TODO: The interactive tutorial is currently VS Code specific, both in terms of features and keyboard shortcuts.
@@ -776,8 +662,108 @@ const register = async (
 
     return {
         disposable: vscode.Disposable.from(...disposables),
-        onConfigurationChange,
     }
+}
+
+// Registers listeners to trigger parsing of visible documents
+function registerParserListeners(disposables: vscode.Disposable[]) {
+    void parseAllVisibleDocuments()
+    disposables.push(vscode.window.onDidChangeVisibleTextEditors(parseAllVisibleDocuments))
+    disposables.push(vscode.workspace.onDidChangeTextDocument(updateParseTreeOnEdit))
+}
+
+interface RegisterChatOptions {
+    context: vscode.ExtensionContext
+    configWatcher: ConfigWatcher<ConfigurationWithAccessToken>
+    platform: PlatformContext
+    chatClient: ChatClient
+    guardrails: Guardrails
+    editor: VSCodeEditor
+    authProvider: AuthProvider
+    contextProvider: ContextProvider
+    enterpriseContextFactory: EnterpriseContextFactory
+    localEmbeddings?: LocalEmbeddingsController
+    contextRanking?: ContextRankingController
+    symfRunner?: SymfRunner
+}
+
+function registerChat(
+    {
+        context,
+        configWatcher,
+        platform,
+        chatClient,
+        guardrails,
+        editor,
+        authProvider,
+        contextProvider,
+        enterpriseContextFactory,
+        localEmbeddings,
+        contextRanking,
+        symfRunner,
+    }: RegisterChatOptions,
+    disposables: vscode.Disposable[]
+): {
+    chatManager: ChatManager
+    editorManager: EditManager
+} {
+    const initialConfig = configWatcher.get()
+
+    // Shared configuration that is required for chat views to send and receive messages
+    const messageProviderOptions: MessageProviderOptions = {
+        chat: chatClient,
+        guardrails,
+        editor,
+        authProvider,
+        contextProvider,
+    }
+    const chatManager = new ChatManager(
+        {
+            ...messageProviderOptions,
+            extensionUri: context.extensionUri,
+            config: initialConfig,
+            startTokenReceiver: platform.startTokenReceiver,
+        },
+        chatClient,
+        enterpriseContextFactory,
+        localEmbeddings || null,
+        contextRanking || null,
+        symfRunner || null,
+        guardrails
+    )
+
+    const ghostHintDecorator = new GhostHintDecorator(authProvider)
+    const editorManager = new EditManager({
+        chat: chatClient,
+        editor,
+        ghostHintDecorator,
+        authProvider,
+        extensionClient: platform.extensionClient,
+    })
+    disposables.push(ghostHintDecorator, editorManager, new CodeActionProvider({ contextProvider }))
+
+    // Register tree views
+    disposables.push(
+        chatManager,
+        vscode.window.registerWebviewViewProvider('cody.chat', chatManager.sidebarViewController, {
+            webviewOptions: { retainContextWhenHidden: true },
+        }),
+
+        // NOTE(beyang): it was necessary to keep this, as it is the vector through which
+        // auth status propagates to the LLM client and guardrails after sign in. Without this,
+        // the auth status does not propagate.
+        contextProvider.configurationChangeEvent.event(async () => {
+            const newConfig = await getFullConfig()
+            configWatcher.set(newConfig)
+        })
+    )
+
+    if (localEmbeddings) {
+        // kick-off embeddings initialization
+        localEmbeddings.start()
+    }
+
+    return { chatManager, editorManager }
 }
 
 /**